VIX-3663 WPF OpenGL Preview
VIX-3663 WPF OpenGL Preview
